requirements.txt
click>=7.0
Flask>=1.1.1
itsdangerous>=1.1.0
Jinja2>=2.10.3
MarkupSafe>=1.1.1
Werkzeug>=0.16.0
psycopg2-binary
